modified exception

+++ cpp/JungMo_Hong/exception/02_stack_unwinding/main.cpp
@@ -41,7 +41,7 @@ void first()
     {
         second();
     }
-    catch (double)
+    catch (int)
     {
         std::cerr << "First caught double exception" << '\n';
     }
